Disable synchronous indexing during import
